+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/SoSly/VillageWorkersPlus/tree/main)
+## [0.1.1-alpha](https://github.com/SoSly/VillageWorkersPlus/releases/tag/0.1.1-alpha)
 
 ### Fixed
 - fixed critical performance issue causing high CPU load